Add files via upload
This report explores the various trends and findings for how the audience and fanbase reacted during the 2025 U.S. Open (Golf).  It was made to help the sales and marketing team make informed decisions so that when a big sponsor does request a report, we (the Data Analytics team) can tell them that we are capable of creating these kinds of reports in the future.